move ContactIcons to icons folder
